refactor: rename vyper/codegen to vyper/codegen_legacy, create vyper/codegen_shared
Rename vyper/codegen/ to vyper/codegen_legacy/ to make room for venom
as the default codegen. Extract shared utilities into vyper/codegen_shared/:

- abi_utils.py: is_tuple_like, needs_external_call_wrap,
  calculate_type_for_external_return, abi_encoding_matches_vyper,
  DYNAMIC_ARRAY_OVERHEAD
- arithmetic.py: calculate_largest_base, calculate_largest_power
- function_info.py: _FuncIRInfo, EntryPointInfo, init_ir_info
- jumptable_utils.py: selector dispatch utilities

codegen_venom now imports from codegen_shared instead of codegen_legacy.
All references to vyper.codegen updated to vyper.codegen_legacy.
