+++ tests/functional/codegen/types/numbers/test_exponents.py
@@ -2,7 +2,7 @@ import pytest
 from hypothesis import example, given, settings
 from hypothesis import strategies as st
 
-from vyper.codegen.arithmetic import calculate_largest_base, calculate_largest_power
+from vyper.codegen_legacy.arithmetic import calculate_largest_base, calculate_largest_power
 from vyper.compiler import compile_code
 from vyper.exceptions import InvalidLiteral
