+++ tests/functional/codegen/types/test_identifier_naming.py
@@ -2,7 +2,7 @@ import pytest
 
 from vyper.ast.identifiers import RESERVED_KEYWORDS
 from vyper.builtins.functions import BUILTIN_FUNCTIONS
-from vyper.codegen.expr import ENVIRONMENT_VARIABLES
+from vyper.codegen_legacy.expr import ENVIRONMENT_VARIABLES
 from vyper.exceptions import NamespaceCollision, StructureException, SyntaxException
 from vyper.semantics.types.primitives import AddressT
