+++ tests/unit/compiler/ir/test_compile_ir.py
@@ -1,6 +1,6 @@
 import pytest
 
-from vyper.codegen.ir_node import IRnode
+from vyper.codegen_legacy.ir_node import IRnode
 from vyper.evm.opcodes import version_check
 from vyper.ir import compile_ir
 from vyper.ir.s_expressions import parse_s_exp
