+++ vyper/cli/vyper_ir.py
@@ -3,7 +3,7 @@ import argparse
 import sys
 
 import vyper
-from vyper.codegen.ir_node import IRnode
+from vyper.codegen_legacy.ir_node import IRnode
 from vyper.ir import compile_ir, optimizer
 from vyper.ir.s_expressions import parse_s_exp
