+++ vyper/codegen_legacy/events.py
@@ -1,7 +1,7 @@
-from vyper.codegen.abi_encoder import abi_encode
-from vyper.codegen.core import get_type_for_exact_size, ir_tuple_from_args, unwrap_location
-from vyper.codegen.ir_node import IRnode
-from vyper.codegen.keccak256_helper import keccak256_helper
+from vyper.codegen_legacy.abi_encoder import abi_encode
+from vyper.codegen_legacy.core import get_type_for_exact_size, ir_tuple_from_args, unwrap_location
+from vyper.codegen_legacy.ir_node import IRnode
+from vyper.codegen_legacy.keccak256_helper import keccak256_helper
 from vyper.exceptions import TypeMismatch
 from vyper.semantics.types.bytestrings import _BytestringT
