+++ vyper/codegen_legacy/jumptable_utils.py
