+++ vyper/codegen_legacy/keccak256_helper.py
@@ -1,7 +1,7 @@
 from math import ceil
 
-from vyper.codegen.core import bytes_data_ptr, ensure_in_memory, get_bytearray_length
-from vyper.codegen.ir_node import IRnode
+from vyper.codegen_legacy.core import bytes_data_ptr, ensure_in_memory, get_bytearray_length
+from vyper.codegen_legacy.ir_node import IRnode
 from vyper.exceptions import CompilerPanic
 from vyper.semantics.types.bytestrings import _BytestringT
 from vyper.semantics.types.shortcuts import BYTES32_T
