+++ vyper/codegen_legacy/module.py
@@ -3,13 +3,14 @@
 from typing import Any, List
 
 import vyper.ast as vy_ast
-from vyper.codegen import core, jumptable_utils
-from vyper.codegen.core import shr
-from vyper.codegen.function_definitions import (
+from vyper.codegen_legacy import core
+from vyper.codegen_shared import jumptable_utils
+from vyper.codegen_legacy.core import shr
+from vyper.codegen_legacy.function_definitions import (
     generate_ir_for_external_function,
     generate_ir_for_internal_function,
 )
-from vyper.codegen.ir_node import IRnode
+from vyper.codegen_legacy.ir_node import IRnode
 from vyper.compiler.settings import _is_debug_mode
 from vyper.exceptions import CompilerPanic
 from vyper.semantics.types.module import ModuleT
