+++ vyper/codegen_shared/__init__.py
@@ -1,0 +1,6 @@
+# Shared utilities between codegen_legacy and codegen_venom.
+#
+# If codegen_legacy is ever removed, these should be refactored —
+# functions here that only exist for legacy compat can be deleted,
+# and the rest can be moved to more specific homes (e.g. vyper/abi.py,
+# vyper/semantics/types/, etc.).
